feat: jgrss/batch val metric logging (#46)
* moved predict callback

* test logger

* test logging

* test

* add name

* test

* test

* test

* test

* test

* test

* test

* test

* test

* test csvlogger

* test

* test

* test

* test

* test

* test

* add CLI arg

* change the number of logs

* update docstring
