+++ setup.cfg
@@ -51,6 +51,7 @@ install_requires =
     torchmetrics>=0.10.0,<0.11.0
     ray>=2.0.0
     gudhi>=3.7.1
+    pyarrow>=11.0.0
     geowombat@git+https://github.com/jgrss/geowombat.git@v2.1.2
     tsaug@git+https://github.com/jgrss/tsaug.git
     geosample@git+https://github.com/jgrss/geosample.git
