+++ src/cultionet/model.py
@@ -2,20 +2,22 @@ import typing as T
 from pathlib import Path
 import logging
 import json
-import filelock
 
 import numpy as np
 
+from .callbacks import LightningGTiffWriter
 from .data.const import SCALE_FACTOR
 from .data.datasets import EdgeDataset, zscores
 from .data.modules import EdgeDataModule
-from .models.lightning import CultioLitModel, MaskRCNNLitModel, TemperatureScaling
+from .models.lightning import (
+    CultioLitModel,
+    MaskRCNNLitModel,
+    TemperatureScaling
+)
 from .utils.reshape import ModelOutputs
 from .utils.logging import set_color_logger
 
-import geowombat as gw
 from scipy.stats import mode as sci_mode
-import rasterio as rio
 from rasterio.windows import Window
 import torch
 from torch_geometric.data import Data
@@ -24,10 +26,10 @@ from pytorch_lightning.callbacks import (
     ModelCheckpoint,
     LearningRateMonitor,
     StochasticWeightAveraging,
-    ModelPruning,
-    BasePredictionWriter
+    ModelPruning
 )
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
+from pytorch_lightning.loggers import CSVLogger, TensorBoardLogger
 from torchvision import transforms
 
 logging.getLogger('lightning').addHandler(logging.NullHandler())
@@ -187,7 +189,7 @@ def fit_maskrcnn(
         devices=1 if device == 'gpu' else None,
         num_processes=0,
         accelerator=device,
-        log_every_n_steps=10,
+        log_every_n_steps=50,
         profiler=None,
         deterministic=False,
         benchmark=False
@@ -249,7 +251,8 @@ def fit(
     stochastic_weight_averaging: T.Optional[bool] = False,
     stochastic_weight_averaging_lr: T.Optional[float] = 0.05,
     stochastic_weight_averaging_start: T.Optional[float] = 0.8,
-    model_pruning: T.Optional[bool] = False
+    model_pruning: T.Optional[bool] = False,
+    save_batch_val_metrics: T.Optional[bool] = False
 ):
     """Fits a model
 
@@ -295,6 +298,7 @@ def fit(
         stochastic_weight_averaging_start (Optional[float]): The stochastic weight averaging epoch start.
             Default is 0.8.
         model_pruning (Optional[bool]): Whether to prune the model. Default is False.
+        save_batch_val_metrics (Optional[bool]): Whether to save batch validation metrics to a parquet file.
     """
     ckpt_file = Path(ckpt_file)
 
@@ -347,7 +351,8 @@ def fit(
         weight_decay=weight_decay,
         class_counts=class_counts,
         edge_class=edge_class,
-        scale_pos_weight=scale_pos_weight
+        scale_pos_weight=scale_pos_weight,
+        save_batch_val_metrics=save_batch_val_metrics
     )
 
     if reset_model:
@@ -416,7 +421,7 @@ def fit(
         devices=1 if device == 'gpu' else None,
         num_processes=0,
         accelerator=device,
-        log_every_n_steps=10,
+        log_every_n_steps=50,
         profiler=profiler,
         deterministic=False,
         benchmark=False
@@ -462,7 +467,7 @@ def fit(
         devices=1 if device == 'gpu' else None,
         num_processes=0,
         accelerator=device,
-        log_every_n_steps=10,
+        log_every_n_steps=50,
         profiler=profiler,
         deterministic=False,
         benchmark=False
@@ -570,152 +575,6 @@ def load_model(
         lit_model.freeze()
 
     return trainer, lit_model
-
-
-class LightningGTiffWriter(BasePredictionWriter):
-    def __init__(
-        self,
-        reference_image: Path,
-        out_path: Path,
-        num_classes: int,
-        ref_res: float,
-        resampling,
-        compression: str,
-        write_interval: str = 'batch'
-    ):
-        super().__init__(write_interval)
-        self.reference_image = reference_image
-        self.out_path = out_path
-        self.out_path.parent.mkdir(parents=True, exist_ok=True)
-
-        with gw.config.update(ref_res=ref_res):
-            with gw.open(reference_image, resampling=resampling) as src:
-                profile = {
-                    'crs': src.crs,
-                    'transform': src.gw.transform,
-                    'height': src.gw.nrows,
-                    'width': src.gw.ncols,
-                    # distance (+1) + edge (+1) + crop (+1) crop types (+N)
-                    # `num_classes` includes background
-                    'count': 3 + num_classes - 1,
-                    'dtype': 'uint16',
-                    'blockxsize': 64 if 64 < src.gw.ncols else src.gw.ncols,
-                    'blockysize': 64 if 64 < src.gw.nrows else src.gw.nrows,
-                    'driver': 'GTiff',
-                    'sharing': False,
-                    'compress': compression
-                }
-        profile['tiled'] = True if max(profile['blockxsize'], profile['blockysize']) >= 16 else False
-        with rio.open(out_path, mode='w', **profile):
-            pass
-        self.dst = rio.open(out_path, mode='r+')
-
-    def write_on_epoch_end(self, trainer, pl_module, predictions, batch_indices):
-        self.dst.close()
-
-    def reshape_predictions(
-        self,
-        batch: Data,
-        distance_batch: torch.Tensor,
-        edge_batch: torch.Tensor,
-        crop_batch: torch.Tensor,
-        crop_type_batch: T.Union[torch.Tensor, None],
-        batch_index: int
-    ) -> T.Tuple[
-        torch.Tensor, torch.Tensor, torch.Tensor, T.Union[torch.Tensor, None]
-    ]:
-        pad_slice2d = (
-            slice(
-                int(batch.row_pad_before[batch_index]),
-                int(batch.height[batch_index])-int(batch.row_pad_after[batch_index])
-            ),
-            slice(
-                int(batch.col_pad_before[batch_index]),
-                int(batch.width[batch_index])-int(batch.col_pad_after[batch_index])
-            )
-        )
-        pad_slice3d = (
-            slice(0, None),
-            slice(
-                int(batch.row_pad_before[batch_index]),
-                int(batch.height[batch_index])-int(batch.row_pad_after[batch_index])
-            ),
-            slice(
-                int(batch.col_pad_before[batch_index]),
-                int(batch.width[batch_index])-int(batch.col_pad_after[batch_index])
-            )
-        )
-        rheight = pad_slice2d[0].stop - pad_slice2d[0].start
-        rwidth = pad_slice2d[1].stop - pad_slice2d[1].start
-        def reshaper(x: torch.Tensor, channel_dims: int) -> torch.Tensor:
-            if channel_dims == 1:
-                return x.reshape(
-                    int(batch.height[batch_index]), int(batch.width[batch_index])
-                )[pad_slice2d].contiguous().view(-1)[:, None]
-            else:
-                return x.t().reshape(
-                    channel_dims,
-                    int(batch.height[batch_index]),
-                    int(batch.width[batch_index])
-                )[pad_slice3d].permute(1, 2, 0).reshape(rheight * rwidth, channel_dims)
-
-        distance_batch = reshaper(distance_batch, channel_dims=1)
-        edge_batch = reshaper(edge_batch, channel_dims=1)
-        crop_batch = reshaper(crop_batch, channel_dims=2)
-        if crop_type_batch is not None:
-            num_classes = crop_type_batch.size(1)
-            crop_type_batch = reshaper(crop_type_batch, channel_dims=num_classes)
-
-        return distance_batch, edge_batch, crop_batch, crop_type_batch
-
-    def write_on_batch_end(
-        self, trainer, pl_module, prediction, batch_indices, batch, batch_idx, dataloader_idx
-    ):
-        distance = prediction['dist']
-        edge = prediction['edge']
-        crop = prediction['crop']
-        crop_type = prediction['crop_type']
-        for batch_index in batch.batch.unique():
-            mask = batch.batch == batch_index
-            w = Window(
-                row_off=int(batch.window_row_off[batch_index]),
-                col_off=int(batch.window_col_off[batch_index]),
-                height=int(batch.window_height[batch_index]),
-                width=int(batch.window_width[batch_index])
-            )
-            w_pad = Window(
-                row_off=int(batch.window_pad_row_off[batch_index]),
-                col_off=int(batch.window_pad_col_off[batch_index]),
-                height=int(batch.window_pad_height[batch_index]),
-                width=int(batch.window_pad_width[batch_index])
-            )
-            distance_batch, edge_batch, crop_batch, crop_type_batch = self.reshape_predictions(
-                batch=batch,
-                distance_batch=distance[mask],
-                edge_batch=edge[mask],
-                crop_batch=crop[mask],
-                crop_type_batch=crop_type[mask] if crop_type is not None else None,
-                batch_index=batch_index
-            )
-            if crop_type_batch is None:
-                crop_type_batch = torch.zeros((crop_batch.size(0), 2), dtype=crop_batch.dtype)
-            mo = ModelOutputs(
-                distance=distance_batch,
-                edge=edge_batch,
-                crop=crop_batch,
-                crop_type=crop_type_batch,
-                instances=None,
-                apply_softmax=False
-            )
-            stack = mo.stack_outputs(w, w_pad)
-            stack = (stack * SCALE_FACTOR).clip(0, SCALE_FACTOR)
-
-            with filelock.FileLock('./dst.lock'):
-                self.dst.write(
-                    stack,
-                    indexes=range(1, self.dst.profile['count']+1),
-                    window=w
-                )
 
 
 def predict_lightning(
