+++ src/cultionet/scripts/cultionet.py
@@ -1247,7 +1247,8 @@ def train_model(args):
         stochastic_weight_averaging=args.stochastic_weight_averaging,
         stochastic_weight_averaging_lr=args.stochastic_weight_averaging_lr,
         stochastic_weight_averaging_start=args.stochastic_weight_averaging_start,
-        model_pruning=args.model_pruning
+        model_pruning=args.model_pruning,
+        save_batch_val_metrics=args.save_batch_val_metrics
     )
